% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilovecats.R
\name{ilovecats}
\alias{ilovecats}
\title{A function to establish who your cat is and if you named them wrong}
\usage{
ilovecats(age, name, price)
}
\arguments{
\item{age}{current age of your cat}

\item{name}{what is your cat's name}

\item{price}{did you pay anything for the cat? adoption fee?}
}
\value{
a sentence describing who is cat
}
\description{
A function to establish who your cat is and if you named them wrong
}
\examples{
ilovecats(1,"shrimp",0)
}
